refactor(cli): add report formatter registry
