+++ packages/cli/src/core/formats/format-json.ts
@@ -1,0 +1,5 @@
+import type { Report } from '../analyze.js';
+
+export function formatJson(report: Report): string {
+  return JSON.stringify(report, null, 2);
+}
