+++ packages/cli/src/core/formats/format-sarif.ts
@@ -1,0 +1,5 @@
+import type { Report } from '../analyze.js';
+
+export function formatSarif(_report: Report): string {
+  throw new Error('SARIF formatting is not yet implemented.');
+}
