+++ packages/cli/src/core/reporters/json.ts
@@ -1,5 +1,6 @@
 import type { Report } from '../analyze.js';
+import { formatJson } from '../formats/format-json.js';
 
 export function createJsonReport(report: Report): string {
-  return JSON.stringify(report, null, 2);
+  return formatJson(report);
 }
